refactor: Move Coroutines related things to the proper package
