+++ app/src/main/java/com/monstarlab/core/coroutines/CoroutinesModule.kt
@@ -1,6 +1,5 @@
-package com.monstarlab.core.injection
+package com.monstarlab.core.coroutines
 
-import com.monstarlab.core.coroutines.CloseableCoroutineScope
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
